backend/requirements.txt
# --- Core API ---
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic==2.9.2
python-multipart==0.0.9

# --- Database (PostgreSQL + PostGIS + AI Vector) ---
sqlalchemy>=2.0.0
psycopg2-binary==2.9.9
geoalchemy2          # 关键：让 SQLAlchemy 支持 PostGIS 地理数据
pgvector             # 关键：让 SQLAlchemy 支持 AI 向量搜索

# --- Geospatial & Trail Processing (HikeBot 核心升级) ---
osmnx                # 关键：下载 OpenStreetMap 步道数据
networkx             # 处理路网拓扑结构
geopandas            # 处理地理数据框 (GeoDataFrame)
shapely              # 处理几何对象 (Point, LineString)
rasterio             # 处理海拔数据 (.tif DEM 文件)
scipy                # osmnx 依赖它进行科学计算

# --- AI & LLM ---
openai>=1.0.0
thefuzz              # 模糊匹配字符串

# --- Data Science & Utils ---
pandas==2.2.3
numpy
python-dateutil
python-dotenv        # 加载 .env 环境变量

# --- Networking & Scraping ---
requests==2.31.0
httpx==0.27.2
beautifulsoup4
lxml

# --- Weather Services ---
openmeteo-requests==1.1.0
requests-cache==1.2.0
retry-requests==2.0.0